Etapa3/requirements.txt
pandas
psycopg2-binary
requests
shapely
geopandas